fix: split type and constant in home page
